Add the ability to build docs quitely
Currently, the makefile in `doc/` is very verbose printing all sorts of progress lines. More importantly, it dumps the content of `doxygen.log` to stdout after the build finishes. This is not always necessary: For example the doxygen travis job prints the content of `doxygen.log` twice.

This commit changes the makefile so that:

* Running `make` will exhibit the same behaviour as it does before this commit, including printing the contents of `doxygen.log`
* Running `make doxygen.log` will exclude all debugging / progress output